+++ rs/tests/consensus/utils/src/lib.rs
@@ -5,7 +5,7 @@ use ic_system_test_driver::driver::test_env_api::set_var_to_path;
 pub mod node;
 pub mod rw_message;
 pub mod ssh_access;
-pub mod subnet_recovery;
+pub mod subnet;
 pub mod upgrade;
 
 /**
