+++ rs/tests/src/orchestrator/rotate_ecdsa_idkg_key.rs
@@ -28,7 +28,7 @@ use std::time::{Duration, SystemTime};
 
 use anyhow::bail;
 use ic_base_types::{NodeId, RegistryVersion};
-use ic_consensus_system_test_utils::subnet_recovery::enable_chain_key_on_subnet;
+use ic_consensus_system_test_utils::subnet::enable_chain_key_on_subnet;
 use ic_consensus_threshold_sig_system_test_utils::make_key_ids_for_all_schemes;
 use ic_consensus_threshold_sig_system_test_utils::run_chain_key_signature_test;
 use ic_interfaces_registry::RegistryValue;
